+++ CHANGELOG.md
@@ -25,10 +25,12 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Fixed
 
+- Comments in the *Fallback DNS servers* field in the UI ([#6182]).
 - Empty or default Safe Browsing and Parental Control settings ([#6181]).
 - Various UI issues.
 
 [#6181]: https://github.com/AdguardTeam/AdGuardHome/issues/6181
+[#6182]: https://github.com/AdguardTeam/AdGuardHome/issues/6182
 
 <!--
 NOTE: Add new changes ABOVE THIS COMMENT.