+++ CHANGELOG.md
@@ -25,12 +25,14 @@ NOTE: Add new changes BELOW THIS COMMENT.
 
 ### Fixed
 
+- Incorrect original answer when a response is filtered ([#6183]).
 - Comments in the *Fallback DNS servers* field in the UI ([#6182]).
 - Empty or default Safe Browsing and Parental Control settings ([#6181]).
 - Various UI issues.
 
 [#6181]: https://github.com/AdguardTeam/AdGuardHome/issues/6181
 [#6182]: https://github.com/AdguardTeam/AdGuardHome/issues/6182
+[#6183]: https://github.com/AdguardTeam/AdGuardHome/issues/6183
 
 <!--
 NOTE: Add new changes ABOVE THIS COMMENT.