+++ CHANGELOG.md
@@ -14,11 +14,11 @@ and this project adheres to
 <!--
 ## [v0.108.0] - TBA
 
-## [v0.107.38] - 2023-09-20 (APPROX.)
+## [v0.107.39] - 2023-09-27 (APPROX.)
 
-See also the [v0.107.38 GitHub milestone][ms-v0.107.38].
+See also the [v0.107.39 GitHub milestone][ms-v0.107.39].
 
-[ms-v0.107.38]: https://github.com/AdguardTeam/AdGuardHome/milestone/73?closed=1
+[ms-v0.107.39]: https://github.com/AdguardTeam/AdGuardHome/milestone/74?closed=1
 
 NOTE: Add new changes BELOW THIS COMMENT.
 -->
@@ -26,6 +26,25 @@ NOTE: Add new changes BELOW THIS COMMENT.
 <!--
 NOTE: Add new changes ABOVE THIS COMMENT.
 -->
+
+
+
+## [v0.107.38] - 2023-09-11
+
+See also the [v0.107.38 GitHub milestone][ms-v0.107.38].
+
+### Fixed
+
+- Incorrect original answer when a response is filtered ([#6183]).
+- Comments in the *Fallback DNS servers* field in the UI ([#6182]).
+- Empty or default Safe Browsing and Parental Control settings ([#6181]).
+- Various UI issues.
+
+[#6181]: https://github.com/AdguardTeam/AdGuardHome/issues/6181
+[#6182]: https://github.com/AdguardTeam/AdGuardHome/issues/6182
+[#6183]: https://github.com/AdguardTeam/AdGuardHome/issues/6183
+
+[ms-v0.107.38]: https://github.com/AdguardTeam/AdGuardHome/milestone/73?closed=1
 
 
 
@@ -218,6 +237,10 @@ See also the [v0.107.36 GitHub milestone][ms-v0.107.36].
   regular expression.
 - Client hostnames not resolving when upstream server responds with zero-TTL
   records ([#6046]).
+
+### Removed
+
+- Go 1.19 support, as it has reached end of life.
 
 [#6046]: https://github.com/AdguardTeam/AdGuardHome/issues/6046
 [#6049]: https://github.com/AdguardTeam/AdGuardHome/issues/6049
@@ -2464,11 +2487,12 @@ See also the [v0.104.2 GitHub milestone][ms-v0.104.2].
 
 
 <!--
-[Unreleased]: https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.38...HEAD
-[v0.107.38]:  https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.37...v0.107.38
+[Unreleased]: https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.39...HEAD
+[v0.107.39]:  https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.38...v0.107.39
 -->
 
-[Unreleased]: https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.37...HEAD
+[Unreleased]: https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.38...HEAD
+[v0.107.38]:  https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.37...v0.107.38
 [v0.107.37]:  https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.36...v0.107.37
 [v0.107.36]:  https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.35...v0.107.36
 [v0.107.35]:  https://github.com/AdguardTeam/AdGuardHome/compare/v0.107.34...v0.107.35